refactor(cobranca) - Atualização para atender padrões do projeto
