+++ src/main/java/com/divertech/raxae/cobranca/infra/DespesaInfraRepository.java
@@ -2,7 +2,9 @@ package com.divertech.raxae.cobranca.infra;
 
 import com.divertech.raxae.cobranca.domain.Despesa;
 import com.divertech.raxae.cobranca.repository.DespesaRepository;
+import com.divertech.raxae.handler.APIException;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Repository;
 
 import java.util.UUID;
@@ -17,8 +19,11 @@ public class DespesaInfraRepository implements DespesaRepository {
     public Despesa salvar(Despesa despesa) {
         return jpaRepository.save(despesa);
     }
+
     @Override
     public Despesa buscaPorId(UUID id) {
-        return jpaRepository.findById(id).orElse(null);
+        return jpaRepository.findById(id)
+                .orElseThrow(() -> APIException.build(HttpStatus.NOT_FOUND, "Despesa não encontrada!"));
     }
+
 }
